refactor: move colony dashboard header to home partials
